Revert "De-JSPM/Babel crosswords"
