+++ static/src/javascripts/es6/projects/common/modules/crosswords/constants.js
@@ -1,0 +1,6 @@
+export default {
+    cellSize: 31,
+    borderSize: 1,
+    numberSize: 9
+};
+
