+++ static/src/javascripts/es6/projects/common/modules/crosswords/keycodes.js
@@ -1,0 +1,10 @@
+export default {
+    backspace: 8,
+    tab: 9,
+    left: 37,
+    up: 38,
+    right: 39,
+    down: 40,
+    'delete': 46
+};
+
