--- static/src/javascripts/projects/common/modules/crosswords/constants.js
@@ -1,7 +1,0 @@
-define(function () {
-    return {
-        cellSize: 31,
-        borderSize: 1,
-        numberSize: 9
-    };
-});
