--- static/src/javascripts/projects/common/modules/crosswords/keycodes.js
@@ -1,11 +1,0 @@
-define(function () {
-    return {
-        backspace: 8,
-        tab: 9,
-        left: 37,
-        up: 38,
-        right: 39,
-        down: 40,
-        'delete': 46
-    };
-});
