+++ static/src/javascripts/test/.eslintrc
@@ -1,0 +1,37 @@
+{
+    "plugins": [
+        "jasmine"
+    ],
+    "env": {
+        "jasmine": true
+    },
+    "ecmaFeatures": {
+        "arrowFunctions": true,
+        "binaryLiterals": true,
+        "blockBindings": true,
+        "classes": true,
+        "defaultParams": true,
+        "destructuring": true,
+        "forOf": true,
+        "generators": true,
+        "modules": true,
+        "objectLiteralComputedProperties": true,
+        "objectLiteralDuplicateProperties": true,
+        "objectLiteralShorthandMethods": true,
+        "objectLiteralShorthandProperties": true,
+        "octalLiterals": true,
+        "regexUFlag": true,
+        "regexYFlag": true,
+        "spread": true,
+        "superInFunctions": true,
+        "templateStrings": true,
+        "unicodeCodePointEscapes": true,
+        "globalReturn": true
+    },
+    "rules": {
+        // Overrides old
+        "camelcase": 2,
+        "no-shadow": 2,
+        "strict": 2
+    }
+}
