+++ static/src/javascripts/test/setup.js
@@ -1,0 +1,15 @@
+// http://jasmine.github.io/2.0/introduction.html#section-43
+jasmine.DEFAULT_TIMEOUT_INTERVAL = 10000;
+
+window.guardian = {
+    config: {
+        switches: { },
+        page: { }
+    }
+};
+
+// Omniture variables expected on the page
+
+/*eslint-disable camelcase*/
+window.s_account = 'guardiangu-network';
+/*eslint-enable camelcase*/
